+++ pkg/telemetry/phonehome/interceptor.go
@@ -6,12 +6,10 @@ import (
 	"strings"
 	"sync"
 
-	"github.com/pkg/errors"
 	erroxGRPC "github.com/stackrox/rox/pkg/errox/grpc"
 	"github.com/stackrox/rox/pkg/grpc/authn"
 	grpcError "github.com/stackrox/rox/pkg/grpc/errors"
 	"github.com/stackrox/rox/pkg/grpc/requestinfo"
-	"github.com/stackrox/rox/pkg/httputil"
 	"google.golang.org/grpc"
 )
 
@@ -29,18 +27,8 @@ var (
 	mux = &sync.Mutex{}
 )
 
-// AddInterceptorFunc appends the custom list of telemetry interceptors with the
-// provided function.
-func (cfg *Config) AddInterceptorFunc(event string, f Interceptor) {
-	mux.Lock()
-	defer mux.Unlock()
-	if cfg.interceptors == nil {
-		cfg.interceptors = make(map[string][]Interceptor, 1)
-	}
-	cfg.interceptors[event] = append(cfg.interceptors[event], f)
-}
-
-func (cfg *Config) track(rp *RequestParams, t Telemeter) {
+func (cfg *Config) track(rp *RequestParams) {
+	id := cfg.HashUserAuthID(rp.UserID)
 	for event, funcs := range cfg.interceptors {
 		props := map[string]any{}
 		ok := true
@@ -50,12 +38,12 @@ func (cfg *Config) track(rp *RequestParams, t Telemeter) {
 			}
 		}
 		if ok {
-			t.Track(event, cfg.HashUserAuthID(rp.UserID), props)
+			cfg.telemeter.Track(event, id, props)
 		}
 	}
 }
 
-func (cfg *Config) getGrpcRequestDetails(ctx context.Context, err error, info *grpc.UnaryServerInfo, req any) *RequestParams {
+func getGrpcRequestDetails(ctx context.Context, err error, info *grpc.UnaryServerInfo, req any) *RequestParams {
 	id, iderr := authn.IdentityFromContext(ctx)
 	if iderr != nil {
 		log.Debug("Cannot identify user from context: ", iderr)
@@ -71,7 +59,7 @@ func (cfg *Config) getGrpcRequestDetails(ctx context.Context, err error, info *g
 	}
 }
 
-func (cfg *Config) getHttpRequestDetails(ctx context.Context, r *http.Request, err error) *RequestParams {
+func getHttpRequestDetails(ctx context.Context, r *http.Request, err error) *RequestParams {
 	id, iderr := authn.IdentityFromContext(ctx)
 	if iderr != nil {
 		log.Debug("Cannot identify user from context: ", iderr)
@@ -84,39 +72,4 @@ func (cfg *Config) getHttpRequestDetails(ctx context.Context, r *http.Request, e
 		Code:      grpcError.ErrToHTTPStatus(err),
 		HttpReq:   r,
 	}
-}
-
-// getGRPCInterceptor returns an API interceptor function for GRPC requests.
-func (cfg *Config) getGRPCInterceptor(t Telemeter) grpc.UnaryServerInterceptor {
-	return func(ctx context.Context, req any, info *grpc.UnaryServerInfo, handler grpc.UnaryHandler) (any, error) {
-		resp, err := handler(ctx, req)
-		rp := cfg.getGrpcRequestDetails(ctx, err, info, req)
-		go cfg.track(rp, t)
-		return resp, err
-	}
-}
-
-func statusCodeToError(code *int) error {
-	if code == nil || *code == http.StatusOK {
-		return nil
-	}
-	return errors.Errorf("%d %s", *code, http.StatusText(*code))
-}
-
-// getHTTPInterceptor returns an API interceptor function for HTTP requests.
-func (cfg *Config) getHTTPInterceptor(t Telemeter) httputil.HTTPInterceptor {
-	return func(handler http.Handler) http.Handler {
-		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			statusTrackingWriter := httputil.NewStatusTrackingWriter(w)
-			handler.ServeHTTP(statusTrackingWriter, r)
-			rp := cfg.getHttpRequestDetails(r.Context(), r, statusCodeToError(statusTrackingWriter.GetStatusCode()))
-			go cfg.track(rp, t)
-		})
-	}
-}
-
-// MakeInterceptors returns a couple of interceptors.
-func (cfg *Config) MakeInterceptors() (grpc.UnaryServerInterceptor, httputil.HTTPInterceptor) {
-	t := cfg.Telemeter()
-	return cfg.getGRPCInterceptor(t), cfg.getHTTPInterceptor(t)
 }
